+++ client/src/components/App.js
@@ -1,10 +1,10 @@
 import React, { useState, useEffect } from 'react';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
 
-import Home from './homepage/Home';
-import Cart from './cartpage/Cart';
-import Item from './itempage/Item';
-import Error from './errorpage/Error';
+import Home from './home-page/Home';
+import Cart from './cart-page/Cart';
+import Item from './item-page/Item';
+import Error from './error-page/Error';
 
 function App() {
   const [bacon, setBacon] = useState(null);
